+++ pkg/news.md
@@ -2,3 +2,5 @@
 
 * LassoShooting.fit: handle of NAs in coefficient vector (Bug 1)
 * rlasso: removed bug in construction of model.matrix and change in predict-method (analog for rlassologit)
+* rlassoIVselectX / rlassoIV: change in man pages (exogenous variables are automatically used as instruments)
+* rlassologit.fit: remove "double scaling"